+++ src/circuit/NewBinaryOperation.java
@@ -30,8 +30,8 @@ public class NewBinaryOperation extends CircuitCommand {
     
     @Override
     public void execute() {
-       x = valueStack.pop();
-       y = valueStack.pop();
+       x = (Double) valueStack.pop();
+       y = (Double) valueStack.pop();
        
        Double result = myFun.evaluate(y, x);
        valueStack.push(result);
